# src/main/resources/META-INF/openapi.yaml
#https://app.swaggerhub.com/help/tutorials/openapi-3-tutorial
openapi: 3.0.0
info:
  description: "REST-Emp System"
  version: "1.0.0"
  title: "REST - Employee System"
servers:
  - url: http://localhost:8080/rest-employeeSystem/api/v1

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  
  responses:
    UnauthorizedAdminError:
      description: request requires HTTP authentication for ADMIN_ROLE
    UnauthorizedUserError:
      description: request requires HTTP authentication for USER_ROLE
    NotFound:
      description: The specified resource was not found
  
  schemas:
  
    Employee:
      description: REST - EmployeeSystem employee
      type: object
      required:
        - id
      properties:
        id:
          description: ID of  entity
          type: integer
        version:
          description: version (OptLock) of entity
          type: integer
        createdDate:
          description: creation timestamp of entity
          type: string
          format: date-time
        updatedDate:
          description: last-update timestamp of entity
          type: string
          format: date-time
        firstName:
          type: string
          description: First Name of the employee
          example: 'John'
        lastName:
          type: string
          description: Last Name of the employee
          example: 'Doe'
        email:
          type: string
          description: Email of the employee
          example: 'John.Doe@gmail.com'
        title:
          type: string
          description: Title of the employee
          example: 'Prof.'
        salary:
          type: number
          format: double
          description: Salary of the employee
          example: 100.0
        # additional for Address/Phone/Project/EmployeeTasks
        
    NewEmployee: 
      type: object
      required:
        - firstName  
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        title:
          type: string
        salary:
          type: number
          format: double
        phones:
          type: array
          items:
            $ref: '#/components/schemas/Phones'
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Projects'
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Tasks'

    Phones:
      type: object
      properties:
        phoneType:
          type: string
        areaCode:
          type: string
        phoneNumber:
          type: string
        department:
          type: string
        provider:
          type: string

    Projects:
      type: object
      properties:
        proj_id:
          type: string
        description:
          type: string
        name:
          type: string        

    Tasks:
      type: object
      properties:
        task_description:
          type: string
        task_done:
          type: string
        task_end_date:
          type: string 
          format: date-time
        task_start:
          type: string
          foramt: date-time
          
    EmployeesList:
        description: list of employees
        type: object
        properties:
          employees:
            type: array
            items:
              $ref: '#/components/schemas/Employee'

security:
  - basicAuth: []
  
paths:
  /employee/testNoAuthConnection:
    get:
      summary: tests a connection that doesn't need authorization

  /employee/testAdminAuthorizationConnection:
    get:
      summary: tests a connection that needs admin authorization             
   
  /employee/testUser1AuthorizationConnection:
    get:
      summary: tests a connection that needs user1 authorization  
  
  /employee:
    get:
      responses:
        401:
           $ref: '#/components/responses/UnauthorizedAdminError'
        200:      
          description: OK
          content:
            application/json:
             schema: 
               $ref: '#/components/schemas/EmployeesList'
    post:
      summary: Creates an employee record
      requestBody:
        content:
          application/json:
           schema: 
             $ref: '#/components/schemas/NewEmployee'
      responses:
        201:
          description: Employee Created          
        401:
          $ref: '#/components/responses/UnauthorizedUserError'
              
  /employee/{id}:
    get:
      summary: Retrieves an employee by its id
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: numeric id of the employee to retrieve
      responses:
        200:
          description: Updated an employee record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        401:
          $ref: '#/components/responses/UnauthorizedUserError'
        404:
          $ref: '#/components/responses/NotFound'  
    delete:
      summary: Deletes an employee by its id
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: numeric id of the employee to retrieve
      responses:
        204:
          description: Deleted an employee
        401:
          $ref: '#/components/responses/UnauthorizedUserError'
        404:
          $ref: '#/components/responses/NotFound'  
    put:
      summary: Updates an employee info by its id 
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: numeric id of the employee to retrieve
      responses:
        200:
          description: OK
        401:
          $ref: '#/components/responses/UnauthorizedUserError'
        404:
          $ref: '#/components/responses/NotFound'
      requestBody:
        content:
          application/json:
           schema: 
             $ref: '#/components/schemas/Employee'                         
  